Server/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy==2.0.35
openpyxl==3.1.2
pandas==2.1.4
PyMySQL==1.1.0
